refactor(tests): move test files to match component file paths
