+++ galasa-ui/src/app/test-runs/[slug]/page.tsx
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: EPL-2.0
  */
 import React from 'react';
-import TestRunDetails from '@/components/test-runs/TestRunDetails';
+import TestRunDetails from '@/components/test-runs/test-run-details/TestRunDetails';
 import NotFound from '@/components/common/NotFound';
 import ErrorPage from '@/app/error/page';
 import { getTranslations } from 'next-intl/server';
