+++ galasa-ui/src/tests/components/test-runs/results/TestRunsTable.test.tsx
@@ -7,7 +7,7 @@ import React from 'react';
 import '@testing-library/jest-dom';
 import { render, screen, waitFor, within } from '@testing-library/react';
 import { fireEvent } from '@testing-library/react';
-import TestRunsTable from '@/components/test-runs/TestRunsTable';
+import TestRunsTable from '@/components/test-runs/results/TestRunsTable';
 import { MAX_RECORDS, RESULTS_TABLE_COLUMNS } from '@/utils/constants/common';
 import { useSearchParams, useRouter } from 'next/navigation';
