+++ galasa-ui/src/tests/components/test-runs/test-run-details/LogTab.test.tsx
@@ -6,7 +6,7 @@
 import React from 'react';
 import { render, screen, fireEvent, waitFor } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import LogTab from '@/components/test-runs/LogTab';
+import LogTab from '@/components/test-runs/test-run-details/LogTab';
 import { handleDownload } from '@/utils/artifacts';
 
 // Mock the utility function
