+++ galasa-ui/src/tests/components/test-runs/test-run-details/TestRunPage.test.tsx
@@ -42,8 +42,6 @@ jest.mock('@/app/test-runs/[slug]/page', () => {
 
 import TestRunPage from '@/app/test-runs/[slug]/page';
 // Note: We keep imports of these for mocking API behavior, but tests will assert against the mock-page div
-import TestRunDetails from '@/components/test-runs/TestRunDetails';
-import NotFound from '@/components/common/NotFound';
 import { createAuthenticatedApiConfiguration } from '@/utils/api';
 import { ResultArchiveStoreAPIApi } from '@/generated/galasaapi';
 import ErrorPage from '@/app/error/page';
